backend/requirements.txt
fastapi
uvicorn
motor
faiss-cpu
networkx
ddgs
crawl4ai
google-genai
python-dotenv
pydantic
requests
beautifulsoup4
